refactor: move most of presentation dir contents into new views and view_models dirs
