+++ lib/src/presentation/views/buttons.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:material_symbols_icons/symbols.dart';
 
-import '../application/exit.dart';
+import '../../application/exit.dart';
 
 /// An [ElevatedButton] whose styling signifies that its action is dangerous in some way.
 class DangerButton extends ElevatedButton {
@@ -88,6 +88,7 @@ class RefreshButton extends StatelessWidget {
   }
 }
 
+// TODO move to a file with tactical widgets
 class FireTorpedoesButton extends StatelessWidget {
   const FireTorpedoesButton({
     super.key,
