+++ lib/src/presentation/views/websocket_viewer.dart
